requirements.txt
langchain
# ipykernel
python-dotenv
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
lxml
langchain-openai
chromadb
sentence_transformers
langchain-huggingface
faiss-cpu
langchain_chroma
beautifulsoup4
streamlit
langchain-ollama
langchain-groq
fastapi
uvicorn
langserve
sse_starlette
